V1.5 - Created data class methods working concurrent with data class from signup and login class.
userCreate() and checkUser() in Data class need to be set up
and a database needs to be introduced into the Data Class.
i will instruct how this needs to be done with the other group members

The navigation pane needs to be changed and the format of the app needs
upgrade. The map tracking feature needs to be introduced into the app.